+++ src/main/java/com/kodilla/sportscentre/services/UserService.java
@@ -1,12 +1,9 @@
 package com.kodilla.sportscentre.services;
 
-import com.kodilla.sportscentre.domain.Card;
-import com.kodilla.sportscentre.domain.CardDto;
 import com.kodilla.sportscentre.domain.User;
-import com.kodilla.sportscentre.domain.UserDto;
-import com.kodilla.sportscentre.exceptions.CardNotFoundException;
+import com.kodilla.sportscentre.domain.UserCreateDto;
+import com.kodilla.sportscentre.domain.UserEditDto;
 import com.kodilla.sportscentre.exceptions.UserNotFoundException;
-import com.kodilla.sportscentre.mappers.CardMapper;
 import com.kodilla.sportscentre.mappers.UserMapper;
 import com.kodilla.sportscentre.repositories.UserRepository;
 import lombok.RequiredArgsConstructor;
@@ -21,22 +18,28 @@ public class UserService {
     private final UserRepository userRepository;
     private final UserMapper userMapper;
 
-    public List<UserDto> getAllUsers() {
+    public List<User> getAllUsers() {
         List<User> userList = userRepository.findAll();
-        return userMapper.mapToUserDtoList(userList);
+        return userList;
     }
 
-    public UserDto getUserById(final Long userId) throws UserNotFoundException {
+    public User getUserById(final Long userId) throws UserNotFoundException {
         User user = userRepository.findById(userId).orElseThrow(UserNotFoundException::new);
-        return userMapper.mapToUserDto(user);
+        return user;
     }
 
-    public UserDto saveUser(final User user) {
-        return userMapper.mapToUserDto(userRepository.save(user));
+    public User createUser(final UserCreateDto userCreateDto) {
+        User user = userMapper.mapToUserFromCreate(userCreateDto);
+        return userRepository.save(user);
+    }
+
+    public User editUser(final UserEditDto userEditDto) {
+        User user = userMapper.mapToUserFromEdit(userEditDto);
+        return userRepository.save(user);
     }
 
     public void deleteUser(final Long userId) throws UserNotFoundException {
-        userRepository.findById(userId).orElseThrow(UserNotFoundException::new);
+        if(!userRepository.existsById(userId)) throw new UserNotFoundException();
         userRepository.deleteById(userId);
     }
 }
